+++ Vault-Coder/src/utopia/vault/coder/model/data/Property.scala
@@ -1,5 +1,6 @@
 package utopia.vault.coder.model.data
 
+import utopia.flow.util.StringExtensions._
 import utopia.vault.coder.model.enumeration.PropertyType
 import utopia.vault.coder.model.scala.code.CodePiece
 import utopia.vault.coder.util.NamingUtils
@@ -13,14 +14,16 @@ object Property
 	  * @param description Description of this property (Default = empty)
 	  * @param useDescription Description on how this property is used (Default = empty)
 	  * @param customDefault Default value passed for this property (empty if no default (default))
+	  * @param customSqlDefault Default value passed for this property in the SQL table creation (empty if no default)
 	  * @param customIndexing User-specified setting whether this property should be indexed.
 	  *                       None if user didn't specify, which results in data type -based indexing (default)
 	  * @return A new property
 	  */
 	def apply(name: Name, dataType: PropertyType, description: String = "", useDescription: String = "",
-	          customDefault: String = "", customIndexing: Option[Boolean] = None): Property =
+	          customDefault: String = "", customSqlDefault: String = "",
+	          customIndexing: Option[Boolean] = None): Property =
 		apply(name, NamingUtils.camelToUnderscore(name.singular), dataType, description, useDescription,
-			customDefault, customIndexing)
+			customDefault, customSqlDefault, customIndexing)
 }
 
 /**
@@ -33,11 +36,13 @@ object Property
   * @param description Description of this property (may be empty)
   * @param useDescription Description on how this property is used (may be empty)
   * @param customDefault Default value passed for this property (empty if no default)
+  * @param customSqlDefault Default value passed for this property in the SQL table creation (empty if no default)
   * @param customIndexing User-specified setting whether this property should be indexed.
   *                       None if user didn't specify, which results in data type -based indexing
   */
 case class Property(name: Name, columnName: String, dataType: PropertyType, description: String,
-                    useDescription: String, customDefault: String, customIndexing: Option[Boolean])
+                    useDescription: String, customDefault: String, customSqlDefault: String,
+                    customIndexing: Option[Boolean])
 {
 	// COMPUTED ----------------------------
 	
@@ -45,6 +50,19 @@ case class Property(name: Name, columnName: String, dataType: PropertyType, desc
 	  * @return Whether there exists an index based on this property
 	  */
 	def isIndexed = customIndexing.getOrElse(dataType.createsIndexByDefault)
+	
+	/**
+	  * @return The default value assigned for this property. Empty if no default is provided.
+	  */
+	def default = customDefault.notEmpty match
+	{
+		case Some(default) => CodePiece(default)
+		case None => dataType.baseDefault
+	}
+	/**
+	  * @return The default value assigned for this property in the SQL document / table creation
+	  */
+	def sqlDefault = customSqlDefault.notEmpty.getOrElse(dataType.baseSqlDefault)
 	
 	/**
 	  * @return Code for this property converted to a value. Expects ValueConversions to be imported.